Username Exist Checker & Vote Button Working
The changes in this commit are the following....

- The button used to vote for each dog is now working and will add 1 to votes for the dog above it, it also displays this in real time
- A user can no longer add two usernames of the same. A real time message will inform the user whether or not the username is available. This prevents dogs from being overwritten.

Known possible bugs:
With this change it now means that if a user wants to add more than one dog then each dog will need its own username. This is not ideal and I need to correct this by possibly allowing users to add more dogs under their username. Which in turn means each dog needs a unique id to allow the same user to add same dog names. If a user can add more than one dog i need to develop a way in which that user can be verified.